Update Rakefile: moved css back to css directory
In the SCSS build, the CSS was compiling in the root directory, and aliases were pointing to them from the /docs/assets/css directory. This was causing problems, so I changed the output directory of the css to the /docs/assets/css directory. Also removed version number from the end - personal preference I guess. Oh and added line number comments to the css output for dev. It'll be stripped from the minified version automatically.